消除编译警告

+++ NowCoder/NC0121_Permutation.h
@@ -47,7 +47,7 @@ public:
     {
         std::vector<std::string> res;
         if (str.size() == 0) return res;
-        permutations(str, 0, str.size(), res);
+        permutations(str, 0, (int)str.size(), res);
         return res;
     }
 
@@ -63,7 +63,7 @@ public:
             if (i != index && str[i] == str[index])
                 continue;
             std::swap(str[i], str[index]);
-            permutations(str, index + 1, str.size(), res);
+            permutations(str, index + 1, (int)str.size(), res);
         }
     }
 };
